Rename files and add documentation.
