+++ run-valgrind.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
+
 valgrind --leak-check=full \
          --show-leak-kinds=all \
          --track-origins=yes \
